fix: solve small QA (#2808)
* fix: fix workspace user tab style


* fix: solve no GNB header issue


* fix: not difficult


---------
